+++ src/main/java/com/devnotfound/talenthub/repository/CrawlerLogRepository.java
@@ -3,5 +3,5 @@ package com.devnotfound.talenthub.repository;
 import com.devnotfound.talenthub.entity.CrawlerLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CrawlerLogRepository extends JpaRepository<CrawlerLog, Long> {
+public interface CrawlerLogRepository extends JpaRepository<CrawlerLog, Integer> {
 }
